fix produtos

+++ Codigo/front/front_moeda_estudantil/lib/domain/nova_vantegem/dto/nova_vantagem_dto.dart
@@ -1,7 +1,7 @@
 class NovaVantagemDTO {
   final String titulo;
   final String descricao;
-  final int valor;
+  final double valor;
   final String email;
 
   NovaVantagemDTO({
